chore: Renamed folder from `wallet` to `payments`
